+++ daemon/start.go
@@ -11,10 +11,10 @@ import (
 	"github.com/docker/docker/api/types/events"
 	"github.com/docker/docker/daemon/container"
 	mobyc8dstore "github.com/docker/docker/daemon/containerd"
+	"github.com/docker/docker/daemon/internal/libcontainerd"
 	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/internal/otelutil"
-	"github.com/docker/docker/libcontainerd"
 	"github.com/pkg/errors"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
